test: scope the dropout tests to this change
Drops test_library_metadata_registry_divergence.py. Its main class documented the metadata/registry divergence, which is deliberately not being fixed, and its xfail(strict=True) would never have flipped. Keeps the one class that belongs here, the end-to-end proof of what the refusal prevents, folded into the dropout test file.

Also corrects two claims in the surviving docstrings. The pollution source is not established: the desktop registerLibraries() sweep blamed earlier has had no callers since Nov 2025, and _write_user_config_delta merges only the delta into the on-disk file. gtn init is the remaining live candidate but is unconfirmed. And the exposure is narrowed rather than removed, since switching to system defaults still drops the project layer.
